+++ crates/board/src/crypto/aead.rs
@@ -60,37 +60,6 @@ where
 
 pub type Array<N> = GenericArray<u8, N>;
 
-pub struct Unsupported<Tag: ArrayLength<u8> + Send> {
-    _never: !,
-    _tag: Tag,
-}
-
-impl<Tag: ArrayLength<u8> + Send> Support<AeadSupport> for Unsupported<Tag> {
-    const SUPPORT: AeadSupport = AeadSupport { no_copy: false, in_place_no_copy: false };
-}
-
-impl<Key, Iv, Tag: ArrayLength<u8> + Send> Api<Key, Iv> for Unsupported<Tag>
-where
-    Key: ArrayLength<u8>,
-    Iv: ArrayLength<u8>,
-{
-    type Tag = Tag;
-
-    fn encrypt(
-        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &mut [u8],
-        _: &mut Array<Self::Tag>,
-    ) -> Result<(), Error> {
-        unreachable!()
-    }
-
-    fn decrypt(
-        _: &Array<Key>, _: &Array<Iv>, _: &[u8], _: Option<&[u8]>, _: &Array<Self::Tag>,
-        _: &mut [u8],
-    ) -> Result<(), Error> {
-        unreachable!()
-    }
-}
-
 #[cfg(feature = "internal-software-crypto-aead")]
 mod software {
     use aead::{AeadCore, AeadInPlace};