+++ crates/board/test.sh
@@ -17,11 +17,13 @@ set -e
 
 . "$(git rev-parse --show-toplevel)"/scripts/test-helper.sh
 
+check_software_crypto
 check_board_api api- --features=std, --all-targets
 
 test_helper
 
 cargo test --features=full-api,std
 cargo check --target=thumbv7em-none-eabi --features=full-api
-cargo check --target=thumbv7em-none-eabi --features=full-api,software-crypto
+cargo check --target=thumbv7em-none-eabi \
+  --features=full-api,internal-software-crypto
 cargo check --target=riscv32imc-unknown-none-elf --features=full-api