+++ crates/runner-nordic/test.sh
@@ -23,7 +23,7 @@ test_helper
 
 cargo check --target=thumbv7em-none-eabi --features=wasm,debug
 cargo check --target=thumbv7em-none-eabi \
-  --features=wasm,debug,software-crypto-sha256
+  --features=wasm,debug,internal-software-crypto
 DEFMT_LOG=trace cargo check --target=thumbv7em-none-eabi --features=wasm,debug
 cargo check --target=thumbv7em-none-eabi --features=wasm,release
 cargo check --target=thumbv7em-none-eabi --features=native,release