+++ crates/scheduler/test.sh
@@ -19,6 +19,7 @@ set -e
 
 ensure_applet
 
+check_software_crypto
 check_applet_api applet-api- --features=wasm,std, --all-targets
 check_board_api board-api- --features=wasm,std, --all-targets